refactor(core): rename files to avoid potential conflicts with exports
