+++ tfhe/src/core_crypto/algorithms/lwe_programmable_bootstrapping/mod.rs
@@ -1,10 +1,10 @@
-pub mod fft128;
-pub mod fft64;
-pub mod ntt64;
+pub mod fft128_pbs;
+pub mod fft64_pbs;
+pub mod ntt64_pbs;
 
-pub use fft128::*;
-pub use fft64::*;
-pub use ntt64::*;
+pub use fft128_pbs::*;
+pub use fft64_pbs::*;
+pub use ntt64_pbs::*;
 
 use crate::core_crypto::algorithms::glwe_encryption::allocate_and_trivially_encrypt_new_glwe_ciphertext;
 use crate::core_crypto::commons::parameters::*;
